requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image processing
Pillow>=10.0.0

# AI Providers
google-generativeai>=0.8.0

# Production server
gunicorn>=21.0.0
